feat: fetchMember in test DB
